refactor: character creator to body namespace
