+++ unity-project/Assets/Scripts/Body/CharacterCreator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6d7e21a932602b346bd16732ad6c01df
+guid: f4ed47aaa1507d84392f72f1bb8c4d6d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
